refactor: rebrand into tosshin_dienen_controller
