+++ .github/workflows/build-and-test.yml
@@ -11,8 +11,7 @@ jobs:
       - name: Checking out
         uses: actions/checkout@v2.3.4
       - name: Building and testing
-        uses: ichiro-its/ros2-ci@v0.4.2
+        uses: ichiro-its/ros2-ci@v1.0.0
         with:
           ros2-distro: foxy
-          apt-packages: git
-          pre-build: git clone https://github.com/threeal/tosshin_interfaces /ws/tosshin_interfaces && git clone https://github.com/ichiro-its/musen /ws/musen
+          external-repos: threeal/tosshin#v0.1.0
