+++ package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>tosshin</name>
-  <version>0.1.0</version>
-  <description>Mobile robot navigation for ROS 2</description>
+  <name>tosshin_dienen_controller</name>
+  <version>0.2.0</version>
+  <description>Dienen's controller implementation of Tosshin mobile robot navigation project</description>
   <maintainer email="alfi.maulana.f@gmail.com">Alfi Maulana</maintainer>
   <license>MIT License</license>
   <buildtool_depend>ament_cmake</buildtool_depend>
